Correct README.md
Running this package creates a `.yml` file specifying that the action should run only on branches called `main` and `ci-*`. However, the instructions in `README.md` suggested pushing a branch called `new-ci`. This changes it to `ci-new`.

It would instead be possible to change the code to match the documentation instead of changing the documentation to match the code, of course.